Use late final for _timer fields
The field in `AnsiProgress`had been non-final so that it could be
initialized in the constructor, the initialization expression needs
access to `this`. With null safety this pattern can be expressed with a
`late final` field.

The field in `VerboseLogger` could have been `final` before the
migration.

Append `-dev` to the version since this can't be published until it is
rolled and validated internally.