Fix frozen spinner
Fixes #60

The `late final` field isn't instantiated until it is read, so without
forcing instantiation in the constructor there is no timer updating the
display. Move initialization back into the constructor.

Other cleanup:
- Remove the `_index` fields since it is otherwise identical to
  `_timer.ticks`.
- Use a Github Action config that more closely matches our other
  packages.
- Replace `homepage` with `repository` in the pubspec.

There are no existing tests for the spinner, and adding one would be a
large task - this code was not written with testability in mind. For now
leave this untested.